Split subscriptions to increase the speed
